Updated Tooltips
Added a custom View when clicking a marker

Now displays all relevant information about the lot

Added a directions button to the tip

Reformatted the app to use a new class called ParkingLot. When reading from a Marker, get the int of the title, and this is the ID of the parking lot. Its stored in a HashMap called lotDictionary, this HashMap will have all known lots and the polling function will use this soon.